+++ imports/ui/layouts/app.js
@@ -1,0 +1,8 @@
+import './app.html';
+
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
+
+Template.app.helpers({});
+
+Template.app.events({});