+++ imports/ui/pages/not-found.js
@@ -1,0 +1,7 @@
+import { Meteor } from 'meteor/meteor';
+
+import './not-found.html';
+
+Template.notFound.helpers({});
+
+Template.notFound.events({});