+++ server/main.js
@@ -1,8 +1,1 @@
-import { Meteor } from 'meteor/meteor';
-
-// Define flemay:less-autoprefixer options
-AUTOPREFIXER_OPTIONS = '{ "browsers": ["Chrome 36", "iOS 7"]}';
-
-Meteor.startup(() => {
-  // code to run on server at startup
-});
+import '/imports/startup/server';